+++ core/OS/Profile.cpp
@@ -6,22 +6,22 @@ namespace OS {
 		json_t *j;
 		j = json_object_get(obj, "id");
 		if(j && json_is_integer(j))
-			ret.id = json_integer_value(j);
+			ret.id = (int)(int)json_integer_value(j);
 
 		j = json_object_get(obj, "userid");
 		if(j && json_is_integer(j)) {
-			ret.userid = json_integer_value(j);
+			ret.userid = (int)json_integer_value(j);
 		}
 		else if(!j) {
 			json_t *user_obj = json_object_get(obj, "user");
 			j = json_object_get(user_obj, "id");
 			if(j)
-				ret.userid = json_integer_value(j);
+				ret.userid = (int)json_integer_value(j);
 		}
 
 		j = json_object_get(obj, "namespaceid");
 		if(j && json_is_integer(j))
-			ret.namespaceid = json_integer_value(j);
+			ret.namespaceid = (int)json_integer_value(j);
 
 		j = json_object_get(obj, "nick");
 		if(j && json_is_string(j))
@@ -45,55 +45,55 @@ namespace OS {
 
 		j = json_object_get(obj, "icquin");
 		if(j && json_is_integer(j))
-			ret.icquin = json_integer_value(j);
+			ret.icquin = (int)json_integer_value(j);
 		else 
 			ret.icquin = 0;
 
 		j = json_object_get(obj, "sex");
 		if(j && json_is_integer(j))
-			ret.sex = json_integer_value(j);
+			ret.sex = (int)json_integer_value(j);
 		else 
 			ret.sex = 0;
 
 		j = json_object_get(obj, "zipcode");
 		if(j && json_is_integer(j))
-			ret.zipcode = json_integer_value(j);
+			ret.zipcode = (int)json_integer_value(j);
 		else 
 			ret.zipcode = 0;
 
 		j = json_object_get(obj, "pic");
 		if(j && json_is_integer(j))
-			ret.pic = json_integer_value(j);
+			ret.pic = (int)json_integer_value(j);
 		else 
 			ret.pic = 0;
 
 		j = json_object_get(obj, "ooc");
 		if(j && json_is_integer(j))
-			ret.ooc = json_integer_value(j);
+			ret.ooc = (int)json_integer_value(j);
 		else 
 			ret.ooc = 0;
 
 		j = json_object_get(obj, "ind");
 		if(j && json_is_integer(j))
-			ret.ind = json_integer_value(j);
+			ret.ind = (int)json_integer_value(j);
 		else 
 			ret.ind = 0;
 
 		j = json_object_get(obj, "mar");
 		if(j && json_is_integer(j))
-			ret.mar = json_integer_value(j);
+			ret.mar = (int)json_integer_value(j);
 		else 
 			ret.mar = 0;
 
 		j = json_object_get(obj, "chc");
 		if(j && json_is_integer(j))
-			ret.chc = json_integer_value(j);
+			ret.chc = (int)json_integer_value(j);
 		else 
 			ret.chc = 0;
 
 		j = json_object_get(obj, "i1");
 		if(j && json_is_integer(j))
-			ret.i1 = json_integer_value(j);
+			ret.i1 = (int)json_integer_value(j);
 		else 
 			ret.i1 = 0;
 
@@ -125,11 +125,11 @@ namespace OS {
 
 		j = json_object_get(obj, "lat");
 		if(j && json_is_real(j))
-			ret.lat = json_real_value(j);
+			ret.lat = (float)json_real_value(j);
 
 		j = json_object_get(obj, "lon");
 		if(j && json_is_real(j))
-			ret.lon = json_real_value(j);
+			ret.lon = (float)json_real_value(j);
 
 		j = json_object_get(obj, "deleted");
 		if(j && json_is_boolean(j))
