+++ core/OS/User.cpp
@@ -6,7 +6,7 @@ namespace OS {
 		json_t *j;
 		j = json_object_get(obj, "id");
 		if(j)
-			ret.id = json_integer_value(j);
+			ret.id = (int)json_integer_value(j);
 
 		j = json_object_get(obj, "email");
 		if(j)
@@ -14,31 +14,31 @@ namespace OS {
 
 		j = json_object_get(obj, "partnercode");
 		if(j)
-			ret.partnercode = json_integer_value(j);
+			ret.partnercode = (int)json_integer_value(j);
 
 		j = json_object_get(obj, "videocard1ram");
 		if(j)
-			ret.videocard_ram[0] = json_integer_value(j);
+			ret.videocard_ram[0] = (int)json_integer_value(j);
 
 		j = json_object_get(obj, "videocard2ram");
 		if(j)
-			ret.videocard_ram[1] = json_integer_value(j);
+			ret.videocard_ram[1] = (int)json_integer_value(j);
 
 		j = json_object_get(obj, "cpuspeed");
 		if(j)
-			ret.cpu_speed = json_integer_value(j);
+			ret.cpu_speed = (int)json_integer_value(j);
 
 		j = json_object_get(obj, "cpubrandid");
 		if(j)
-			ret.cpu_brandid = json_integer_value(j);
+			ret.cpu_brandid = (int)json_integer_value(j);
 
 		j = json_object_get(obj, "connectionspeed");
 		if(j)
-			ret.connectionspeed = json_integer_value(j);
+			ret.connectionspeed = (int)json_integer_value(j);
 
 		j = json_object_get(obj, "connectionid");
 		if(j)
-			ret.connectionid = json_integer_value(j);
+			ret.connectionid = (int)json_integer_value(j);
 
 		j = json_object_get(obj, "hasnetwork");
 		if(j)
@@ -46,7 +46,7 @@ namespace OS {
 
 		j = json_object_get(obj, "publicmask");
 		if(j && json_is_integer(j))
-			ret.publicmask = json_integer_value(j);
+			ret.publicmask = (int)json_integer_value(j);
 		else 
 			ret.publicmask = 0;
